Factor out CatchResult
