+++ persistence/src/vespa/persistence/spi/persistenceprovider.cpp
@@ -1,30 +1,12 @@
 // Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
 #include "persistenceprovider.h"
+#include "catchresult.h"
 #include <future>
-#include <cassert>
 
 namespace storage::spi {
 
 PersistenceProvider::~PersistenceProvider() = default;
-
-class CatchResult : public OperationComplete {
-public:
-    CatchResult() : _promisedResult(), _resulthandler(nullptr) {}
-    std::future<Result::UP> future_result() {
-        return _promisedResult.get_future();
-    }
-    void onComplete(Result::UP result) override {
-        _promisedResult.set_value(std::move(result));
-    }
-    void addResultHandler(const ResultHandler * resultHandler) override {
-        assert(_resulthandler == nullptr);
-        _resulthandler = resultHandler;
-    }
-private:
-    std::promise<Result::UP>  _promisedResult;
-    const ResultHandler      *_resulthandler;
-};
 
 Result
 PersistenceProvider::put(const Bucket& bucket, Timestamp timestamp, DocumentSP doc, Context& context) {
